Doc link to Symlink Race Condition Protection updated for EA4.
Case EA-5542:

In the WHM Security Advisor page, the advisory concerning the Bluehost
Symlink protection contains a go link pointing back to the EA3 pages
describing the Symlink Race Condition Protection.

This documentation is updated for EA4, and a new web page exists to
describe the Symlink Race Condition Protection as it applies to EA4.
This commit udpates the go link (for EA4 systems only) to point to the
matching EA4 go link.