Fix symlink race condition links to point to EA4, if appropriate.
Case CPANEL-10169:

In a previous fix, we added an EA4 specific go-link for the
symlink race condition description page, but we only added it for
the Bluehost symlink patch advisory.  This commit updates all the other
advisories to also use the EA4 page, if the system is an EA4 system.