Properly detect sysctl setting on grsec kernel
Case CPANEL-7071: On a grsec kernel, if the sysctl option was not
enabled, cPanel would indicate that the setting was incorrectly set.
This occurred because the word "unknown" was being printed on standard
error, but we only looked for it on standard output.  Capture the error
output as well, so we correctly detect when the option is not enabled.