Promote KernelCare if not installed on cPanel 11.64 and greater.
Case SWAT-22, SWAT-320: Added a check to see if KC is installed, if not
provide an "info" advisory with a link to the KernelCare site for
buying/installing.

The decision whether or not to show the info advisory is based on three
criteria:
* If KC is already installed, don't bother.
* If we're in an environment that *can't* use KC, don't bother.
* If the partner for this server has decreed that "thou shalt not,"
  then don't. If they've listed an alternative path for that--via
  another website or email address, show it, but use that link instead.

SWAT-320: A couple of changes were suggested in review.  Minor
matters..Found a few minor perltidy issues.

Link to install page if KC license is found, but not installed.

Case SWAT-359: Check to see if there is a valid KernelCare license
and and the KernelCare RPM is not installed. If this is the case,
preset a link to the install handler in WHM (the same page that
handles after purchase.) Also added to check to ignore KC purchase
check if it's cPanel older than 11.64.

Contributors:

* D Ruth Bavousett <ruth@cpanel.net>
* Brett Estrade <brett@cpanel.net>